Database and Adapter Modified
