+++ app/src/main/java/com/himanshurawat/notes/db/entity/NoteEntity.kt
@@ -16,4 +16,4 @@ data class NoteEntity(
         @ColumnInfo(name = "description")
         var description:String = "",
         @ColumnInfo(name = "date")
-        var date: String = "")
+        var date: Long = 0)
